fix(db): DATE vs TIMESTAMP + Capacity NUMERIC — tests 10→11 passed, manufactura 4/4

+++ web/api/sqlweb-pg/includes/sp/usp_mfg.sql
@@ -215,7 +215,7 @@ RETURNS TABLE(
     "WorkCenterName" VARCHAR,
     "WarehouseId" BIGINT,
     "CostPerHour"    NUMERIC,
-    "Capacity"       INT,
+    "Capacity" NUMERIC,
     "CapacityUom"    VARCHAR,
     "IsActive"       BOOLEAN,
     "CreatedAt"      TIMESTAMP,
